AMP-10315
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE@15603 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/repository/widget/view/indicatorSectorRegionManager.jsp
@@ -133,7 +133,7 @@
 							</a>
                                                         
 							|&nbsp;
-							<a href="/widget/indSectRegManager.do~actType=delete~indSectId=${indSecReg.id}">
+							<a href="/widget/indSectRegManager.do~actType=delete~indSectId=${indSecReg.id}" onclick="return confirm('<digi:trn>Do you want to delete this indicator?</digi:trn>')">
 								<img border="0" src='<digi:file src="images/deleteIcon.gif"/>'>
 							</a>
 						</td>
